requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
bcrypt
python-jose[cryptography]
